// front/src/app/model/keys-storage.enum.ts
export enum KeysStorageEnum {
    PE_CONNECT_INDIVIDU = 'peactions.peConnectIndividu',
    PE_CONNECT_PAYLOAD_STORAGE_SESSION_KEY = 'peactions.peConnectPayload',
    PE_CONNECT_USER_BADGE = 'userBadge',
    INDIVIDU_CONNECTE_STORAGE_SESSION_KEY = 'peactions.demandeurEmploiConnecte',
    DEMANDEUR_EMPLOI_CONNECTE_ROUTE_ACTIVATED_STORAGE_SESSION_KEY = 'peactions.demandeurEmploiConnecte.routeActivated',
    DEMANDEUR_EMPLOI_MESSAGE_NON_AUTORISE = 'peactions.demandeurEmploiConnecte.messageNonAutorisee'
}